Use "position: absolute" on the hover tooltip
Without this, it occupied in-layout flow space for the navigation
bar in the Vector skin (possibly other skins as well).

As a result of this, it pushed away the right-hand tabs and search
bar causing them to wrap to the next line, instead of letting
the tooltip overlap as intended.

It seems this only affected Firefox.